requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
cached-property==1.5.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
djangorestframework==3.10.2
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
idna==2.7
jsonschema==2.6.0
paramiko==2.6.0
postgres==2.2.2
psycopg2-binary==2.8.3
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.2
requests==2.20.1
six==1.12.0
sqlparse==0.3.0
texttable==0.9.1
urllib3==1.24.3
websocket-client==0.56.0
django-money
Pillow
django-rest-auth
django-location-field